% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRT4.R
\name{CRT4}
\alias{CRT4}
\title{Automatic coding for Cognitive Reflection Test 4-item version (Toplak et al., 2014) open-ended responses}
\usage{
CRT4(
  item1 = NULL,
  item2 = NULL,
  item3 = NULL,
  item4 = NULL,
  codingscheme = "categ"
)
}
\arguments{
\item{item1}{Vector of responses to the first CRT question, or NULL if not provided.}

\item{item2}{Vector of responses to the second CRT question, or NULL if not provided.}

\item{item3}{Vector of responses to the third CRT question, or NULL if not provided.}

\item{item4}{Vector of responses to the fourth CRT question, or NULL if not provided.}

\item{codingscheme}{A character string indicating the desired coding scheme. Options are "categ" for the original 1, 2, 3 coding, "sum" for a sum of binary-coded correct answers, or "mean" for an average of binary-coded correct answers. The default is "categ".}
}
\value{
A list containing the coded and, if applicable, binary-coded responses for each provided CRT question. For "sum" or "mean" coding schemes, additional vectors representing these aggregate scores are included.
}
\description{
Applies coding logic to any number of provided CRT question responses and supports multiple coding schemes. This function can output original coded responses, binary-coded responses, and aggregate scores based on these binary codings.
}
\note{
Developed by Giuseppe Corbelli, email: giuseppe.corbelli@uninettunouniversity.net
}
\examples{
# Automated scoring for CRT4 responses using the categorical coding scheme:
reflectR::CRT4(
item1 = c("four", "nineeee", "maybe 4?"),
item2 = c("29", "thirty", "30"),
item3 = c("twentyyyy", "ten I think", "dunno"),
item4 = c("your behind", "poorer", "richer"),
codingscheme = "categ")

# Compute the sum score for CRT4 responses based on binary-coded correctness:
reflectR::CRT4(
item1 = c("four", "nineeee", "maybe 4?"),
item2 = c("29", "thirty", "30"),
item3 = c("twentyyyy", "ten I think", "dunno"),
item4 = c("your behind", "poorer", "richer"),
codingscheme = "sum")$crt_sum
}
