% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRTcoder.R
\name{CRTcoder}
\alias{CRTcoder}
\title{Automatic coding for Cognitive Reflection Test 2 open-ended responses}
\usage{
CRTcoder(
  crt1 = NULL,
  crt2 = NULL,
  crt3 = NULL,
  crt4 = NULL,
  codingscheme = "categorical"
)
}
\arguments{
\item{crt1}{Vector of responses to the first CRT question, or NULL if not provided.}

\item{crt2}{Vector of responses to the second CRT question, or NULL if not provided.}

\item{crt3}{Vector of responses to the third CRT question, or NULL if not provided.}

\item{crt4}{Vector of responses to the fourth CRT question, or NULL if not provided.}

\item{codingscheme}{A character string indicating the desired coding scheme. Options are "categorical" for the original 1, 2, 3 coding, "sum" for a sum of binary-coded correct answers, or "mean" for an average of binary-coded correct answers. The default is "categorical".}
}
\value{
A list containing the coded and, if applicable, binary-coded responses for each provided CRT question. For "sum" or "mean" coding schemes, additional vectors representing these aggregate scores are included.
}
\description{
Applies coding logic to any number of provided CRT question responses and supports multiple coding schemes. This function can output original coded responses, binary-coded responses, and aggregate scores based on these binary codings.
}
\examples{
reflectR::CRTcoder(
  crt1 = c("al primo", "secondo", "1"),
  crt2 = c("7", "otto", "sette"),
  crt3 = c("primo", "carlo", "si chiama primo"),
  crt4 = c("nulla", "27 metri cubi", "zero"),
  codingscheme = "mean"
)
}
