% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRTcoder1.R
\name{CRTcoder1}
\alias{CRTcoder1}
\title{Automatic coding for Cognitive Reflection Test 2 open-ended response n.1}
\usage{
CRTcoder1(risposta)
}
\arguments{
\item{risposta}{The textual response to the CRT-2 first question, as a character string.}
}
\value{
An integer code representing the categorization of the response: 1 for correct, 2 for impulsive but incorrect, and 3 for other responses.
}
\description{
This function automatically codes responses to the first question of the Cognitive Reflection Test 2 (CRT-2), based on predefined regex patterns. It categorizes responses as: 2 for intuitive but incorrect (e.g., "first"), 1 for correct (e.g., "second"), and 3 for any other response.
}
\examples{
reflectR::CRTcoder1(c("al primo", "secondo", "1"))
}
