% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRTcoder2.R
\name{CRTcoder2}
\alias{CRTcoder2}
\title{Automatic coding for Cognitive Reflection Test 2 open-ended response n.2}
\usage{
CRTcoder2(risposta)
}
\arguments{
\item{risposta}{The textual response to the CRT-2 second question, as a character string.}
}
\value{
An integer code representing the categorization of the response: 1 for correct (representing "eight"), 2 for impulsive but incorrect (representing "seven"), and 3 for other types of responses.
}
\description{
Automatically codes responses to the second question of the Cognitive Reflection Test 2 (CRT-2) using predefined regex patterns. It categorizes responses into: 2 for intuitive but incorrect answers related to "7", 1 for the correct answer "8", and 3 for any other responses.
}
\examples{
reflectR::CRTcoder2(c("7", "otto", "sette"))
}
