% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRTcoder3.R
\name{CRTcoder3}
\alias{CRTcoder3}
\title{Automatic coding for Cognitive Reflection Test 2 open-ended response n.3}
\usage{
CRTcoder3(risposta)
}
\arguments{
\item{risposta}{The textual response to the CRT-2 third question, as a character string.}
}
\value{
An integer code representing the categorization of the response: 1 for correct (indicating "carlo"), 2 for impulsive but incorrect (e.g., "primo" or similar), and 3 for other responses.
}
\description{
Automatically codes responses to the third question of the Cognitive Reflection Test 2 (CRT-2) based on predefined regex patterns. It categorizes responses into: 2 for the common intuitive but incorrect answer (e.g., "primo"), 1 for the correct answer ("carlo"), and 3 for any other type of response.
}
\examples{
reflectR::CRTcoder3(c("primo", "carlo", "si chiama primo"))
}
