% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRTcoder4.R
\name{CRTcoder4}
\alias{CRTcoder4}
\title{Automatic coding for Cognitive Reflection Test 2 open-ended response n.4}
\usage{
CRTcoder4(risposta)
}
\arguments{
\item{risposta}{The textual response to the CRT-2 fourth question, as a character string.}
}
\value{
An integer code representing the categorization of the response: 1 for correct (specifically "zero"), 2 for impulsive but incorrect numerical answers, and 3 for other responses.
}
\description{
Automatically codes responses to the fourth question of the Cognitive Reflection Test 2 (CRT-2), distinguishing between intuitive but incorrect numerical responses and the correct response "zero". It categorizes responses as: 2 for any calculation or non-zero number, 1 for the correct answer "zero", and 3 for any other type of response.
}
\examples{
reflectR::CRTcoder4(c("nulla", "27 metri cubi", "zero"))
}
