% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflectR-package.R
\docType{package}
\name{reflectR}
\alias{reflectR}
\alias{reflectR-package}
\title{Automatic Scoring of the Cognitive Reflection Test}
\description{
The reflectR package provides automated coding tools for open-ended responses
to various forms of the Cognitive Reflection Test (CRT). This package aims to
standardize the analysis of CRT responses across broad datasets in cognitive
psychology, decision-making, and related fields, thereby reducing manual effort
and the variability in interpretation of responses.
}
\details{
The package includes functions to code responses from:
\itemize{
  \item \code{\link{CRT}}: Original CRT version by Frederick (2005)
  \item \code{\link{CRT4}}: 4-item CRT version by Toplak et al. (2014)
  \item \code{\link{CRT7}}: 7-item expanded CRT version by Toplak et al. (2014)
  \item \code{\link{CRTlong}}: CRT LONG version by Primi et al. (2016)
  \item \code{\link{CRTtwo}}: CRT version 2 by Thomson & Oppenheimer (2016)
  \item \code{\link{itaCRTtwo}}: Italian version of CRT version 2 by Thomson & Oppenheimer (2016)
}
}
\section{Note}{

While reflectR draws inspiration from the scientific literature on the CRT,
it has been independently developed and does not hold any affiliation with
the original authors of these tests.
}

\section{Acknowledgments}{

Development of this package was significantly enhanced by insights and suggestions
from Dr. Keela Thomson.
Special thanks are also due to Dr. Paolo Giovanni Cicirelli,
Prof. Marinella Paciello, Dr. Carmela Sportelli, and Prof. Francesca D'Errico
for their contributions to the conceptualization and relevance of this tool in
the research domains of personality, social, and cognitive psychology.
}

\references{
Frederick, S. (2005). Cognitive Reflection and Decision Making. \emph{Journal of Economic Perspectives}, 19(4), 25-42.

Toplak, M. E., West, R. F., & Stanovich, K. E. (2014). Assessing miserly information processing: An expansion of the Cognitive Reflection Test. \emph{Thinking & Reasoning}, 20(2), 147-168.

Primi, C., Morsanyi, K., Chiesi, F., Donati, M. A., & Hamilton, J. (2016). The development and testing of a new version of the Cognitive Reflection Test applying item response theory (IRT). \emph{Journal of Behavioral Decision Making}, 29(5), 453-469.

Thomson, K. S., & Oppenheimer, D. M. (2016). Investigating an alternate form of the cognitive reflection test. \emph{Judgment and Decision Making}, 11(1), 99-113.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/g-corbelli/reflectR}
  \item Report bugs at \url{https://github.com/g-corbelli/reflectR/issues}
}

}
\author{
\strong{Maintainer}: Giuseppe Corbelli \email{giuseppe.corbelli@uninettunouniversity.net} (ORCID: 0000-0002-2864-3548)

}
