% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itaCRTtwo.R
\name{itaCRTtwo}
\alias{itaCRTtwo}
\title{Automatic coding for Cognitive Reflection Test 2 open-ended responses (Italian language)}
\usage{
itaCRTtwo(
  item1 = NULL,
  item2 = NULL,
  item3 = NULL,
  item4 = NULL,
  codingscheme = "categ"
)
}
\arguments{
\item{item1}{Vector of responses to the first CRT question, or NULL if not provided.}

\item{item2}{Vector of responses to the second CRT question, or NULL if not provided.}

\item{item3}{Vector of responses to the third CRT question, or NULL if not provided.}

\item{item4}{Vector of responses to the fourth CRT question, or NULL if not provided.}

\item{codingscheme}{A character string indicating the desired coding scheme. Options are "categorical" for the original 1, 2, 3 coding, "sum" for a sum of binary-coded correct answers, or "mean" for an average of binary-coded correct answers. The default is "categorical".}
}
\value{
A list containing the coded and, if applicable, binary-coded responses for each provided CRT question. For "sum" or "mean" coding schemes, additional vectors representing these aggregate scores are included.
}
\description{
Applies coding logic to any number of provided CRT question responses and supports multiple coding schemes. This function can output original coded responses, binary-coded responses, and aggregate scores based on these binary codings.
}
\examples{
reflectR::CRTtwo(
  item1 = c("al primo", "secondo", "1"),
  item2 = c("7", "otto", "sette"),
  item3 = c("primo", "carlo", "si chiama primo"),
  item4 = c("nulla", "27 metri cubi", "zero"),
  codingscheme = "mean"
)
}
